Fixed a bug with displaying/downloading files, and add role permission on orgs

+++ convex/files.ts
@@ -28,7 +28,7 @@ async function hasAccessToOrg(ctx: QueryCtx | MutationCtx, orgId: string) {
 
     if(!user) return null
 
-    const hasAccess = user.orgIds.includes(orgId) || user.tokenIdentifier.includes(orgId)
+    const hasAccess = user.orgIds.some(item => item.orgId === orgId) || user.tokenIdentifier.includes(orgId)
 
     if(!hasAccess) return null
 
@@ -111,6 +111,12 @@ export const deleteFile = mutation({
 
         if(!access){
             throw new ConvexError("no access to file")
+        }
+
+        const isAdmin = access.user.orgIds.find(org => org.orgId === access.file.orgId)?.role === "admin"
+
+        if(!isAdmin){
+            throw new ConvexError("you have no access to delete")
         }
 
         await ctx.db.delete(args.fileId)
